pyproject.toml
[project]
name = "gd-price-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "natsort>=8.4.0",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
